Refactoring and code quality actions.
