+++ src/Cuemon.Core/Diagnostics/ProcessInfo.cs
@@ -30,9 +30,9 @@ namespace Cuemon.Diagnostics
                 builder.Append(Alphanumeric.CaretChar);
                 builder.Append(FormattableString.Invariant($"ThreadCount: {Process.Threads.Count}"));
                 builder.Append(Alphanumeric.CaretChar);
-                builder.Append(FormattableString.Invariant($"WorkingSet64: {ByteMultipleTable.FromBytes(Process.WorkingSet64, o => o.Prefix = UnitPrefix.Decimal).ToString()}"));
+                builder.Append(FormattableString.Invariant($"WorkingSet64: {ByteStorageCapacity.FromBytes(Process.WorkingSet64, o => o.Prefix = UnitPrefix.Decimal)}"));
                 builder.Append(Alphanumeric.CaretChar);
-                builder.Append(FormattableString.Invariant($"TotalWorkingSet64: {ByteMultipleTable.FromBytes(Process.GetProcesses().Select(p => p.WorkingSet64).Sum(), o => o.Prefix = UnitPrefix.Decimal).ToString()}"));
+                builder.Append(FormattableString.Invariant($"TotalWorkingSet64: {ByteStorageCapacity.FromBytes(Process.GetProcesses().Select(p => p.WorkingSet64).Sum(), o => o.Prefix = UnitPrefix.Decimal)}"));
                 builder.Append(Alphanumeric.CaretChar);
                 builder.Append(FormattableString.Invariant($"TotalProcessorTime: {Process.TotalProcessorTime.ToString("G", CultureInfo.InvariantCulture)}"));
             }
