+++ src/Cuemon.Core/ZeroPrefix.cs
@@ -1,0 +1,9 @@
+﻿namespace Cuemon
+{
+    internal sealed class ZeroPrefix : PrefixMultiple
+    {
+        internal ZeroPrefix() : base("", "", 0, 0)
+        {
+        }
+    }
+}
